Move contacts and notifications to user-profile
