+++ app/app.js
@@ -28,8 +28,6 @@ require('./common/common-module.js');
 require('./post/post-module.js');
 require('./setting/setting-module.js');
 require('./set/set-module.js');
-require('./notification/notification-module.js');
-require('./contact/contact-module.js');
 require('./user-profile/user-profile-module.js');
 
 // Make sure we have a window.ushahidi object
@@ -75,8 +73,6 @@ angular.module('app',
         'ushahidi.posts',
         'ushahidi.tools',
         'ushahidi.sets',
-        'ushahidi.notifications',
-        'ushahidi.contacts',
         'ushahidi.user-profile'
     ])
