--- app/contact/contact-module.js
@@ -1,4 +1,0 @@
-angular.module('ushahidi.contacts', [])
-
-.service('ContactEndpoint', require('./services/endpoints/contact.js'));
-
