--- app/notification/notification-module.js
@@ -1,4 +1,0 @@
-angular.module('ushahidi.notifications', [])
-.config(require('./notification-routes.js'))
-.service('NotificationEndpoint', require('./services/endpoints/notification.js'));
-
