+++ app/user-profile/services/endpoints/contact.js
@@ -14,16 +14,12 @@ function (
         get: {
             method: 'GET',
             isArray: true,
-            transformResponse: function (data /*, header*/) {
+            transformResponse: function (data) {
                 return angular.fromJson(data).results;
             }
         },
         update: {
             method: 'PUT'
-        },
-        addContact: {
-            method: 'POST',
-            url: Util.apiUrl('/contacts')
         }
     });
